+++ rules/core/constants.js
@@ -1,7 +1,9 @@
 'use strict';
 
 var COMPOSITION_METHODS = ['compose', 'flow', 'flowRight', 'pipe'];
+var FOREACH_METHODS = ['forEach', 'forEachRight', 'each', 'eachRight', 'forIn', 'forInRight', 'forOwn', 'forOwnRight'];
 
 module.exports = {
-  COMPOSITION_METHODS: COMPOSITION_METHODS
+  COMPOSITION_METHODS: COMPOSITION_METHODS,
+  FOREACH_METHODS: FOREACH_METHODS
 };